Multiple Changes:
Normalized all audio in the game, everything should be the same volume now.

A new mafioso is selected when the godfather dies, a new mafioso is selected when the mafioso dies.

Changed role distribution to be based on role type now. This was a very poorly made update and should be changed in the future.

Signed-off-by: Alex Bange <aebange@gmail.com>